requirements.txt
youtube-transcript-api==0.6.2
yt-dlp
openai-whisper
requests
python-dotenv
